Added support to download datasets as csv, json or xml
It's a little hackey but I did it this way to support downloading
possibly huge data sets without running out of memory (uses about 10MB)
to export around 40k entries.

My local install with a limit of 128MB runs out of memory if we process
this the normal way (build an array of 40k results) and generate the
view. And I didn't want to impose a record limit for exports.

If someone has a better approach, by all means fix it :).